delete debug comment

+++ engine/lib_download.sh
@@ -3,15 +3,15 @@
 FILE_SERVER="http://192.168.2.119/matrix/libs"
 LIBS_TAR="engine_libs.tar"
 
-#if [ ! -d lib ]; then
-#    mkdir "lib"
-#fi
+if [ ! -d lib ]; then
+    mkdir "lib"
+fi
 
-#if [ "`ls -A lib`" = "" ]; then
+if [ "`ls -A lib`" = "" ]; then
     wget -O ./lib/$LIBS_TAR $FILE_SERVER/$LIBS_TAR
     tar -xvf ./lib/$LIBS_TAR -C ./lib
     rm -rf ./lib/$LIBS_TAR
     echo "Dependency library OK."
-#else
-#    echo "Dependency library already OK."
-#fi
+else
+    echo "Dependency library already OK."
+fi
